backend/requirements.txt
Django>=4.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
PyJWT>=2.8.0
django-filter>=23.0
openpyxl>=3.1.0
reportlab>=4.0.0
